+++ src/components/notes/sider/NoteSider.js
@@ -18,6 +18,7 @@ function NoteSider() {
     const noteFilterApi = useNoteFilterApi();
     const tagApi = useTagApi();
 
+    const [searchValue, setSearchValue] = useState('');
     const [openKeys, setOpenKeys] = useState(['general']);
 
     const onSelect = event => {
@@ -143,11 +144,19 @@ function NoteSider() {
                         text={(
                             <Tooltip title="Press enter to search" placement="bottom">
                                 <Input.Search
+                                    value={searchValue}
                                     allowClear={true}
                                     size="small"
                                     placeholder="Search for ..."
                                     style={{ width: '80%' }}
-                                    onSearch={value => onSearch(value)} />
+                                    onChange={event => setSearchValue(event.target.value)}
+                                    onSearch={value => onSearch(value)}
+                                    onKeyDown={event => {
+                                        if (event.key === 'Escape') {
+                                            setSearchValue('');
+                                            onSearch('');
+                                        }
+                                    }} />
                             </Tooltip>
                         )} />
                 </Menu.Item>